moved instance segmentation methods from "segmentation" to "patch_VAE"
