+++ pipeline/patch_VAE.py
@@ -6,11 +6,15 @@ import numpy as np
 import matplotlib.pyplot as plt
 import importlib
 import inspect
+import logging
+log = logging.getLogger(__name__)
+
 from configs.config_reader import YamlReader
 from torch.utils.data import TensorDataset
 
 from SingleCellPatch.extract_patches import process_site_extract_patches, im_adjust
 from SingleCellPatch.generate_trajectories import process_site_build_trajectory, process_well_generate_trajectory_relations
+from SingleCellPatch.instance_clustering import process_site_instance_segmentation
 
 from pipeline.train_utils import zscore, zscore_patch
 import HiddenStateExtractor.vae as vae
@@ -18,6 +22,61 @@ import HiddenStateExtractor.resnet as resnet
 from HiddenStateExtractor.vq_vae_supp import prepare_dataset_v2, vae_preprocess
 
 NETWORK_MODULE = 'run_training'
+
+
+def instance_segmentation(raw_folder: str,
+                          supp_folder: str,
+                          # val_folder: str,
+                          sites: list,
+                          config_: YamlReader,
+                          rerun=False,
+
+                          **kwargs):
+    """ Helper function for instance segmentation
+
+    Wrapper method `process_site_instance_segmentation` will be called, which
+    loads "*_NNProbabilities.npy" files and performs instance segmentation.
+
+    Results will be saved in the supplementary data folder, including:
+        "cell_positions.pkl": dict of cells in each frame (IDs and positions);
+        "cell_pixel_assignments.pkl": dict of pixel compositions of cells
+            in each frame;
+        "segmentation_*.png": image of instance segmentation results.
+
+    Args:
+        raw_folder (str): folder for raw data, segmentation and summarized results
+        supp_folder (str): folder for supplementary data
+        sites (list of str): list of site names
+        config (YamlReader):
+
+    """
+
+    for site in sites:
+        site_path = os.path.join(raw_folder, '%s.npy' % site)
+        site_segmentation_path = os.path.join(raw_folder,
+                                              '%s_NNProbabilities.npy' % site)
+        if not os.path.exists(site_path) or not os.path.exists(site_segmentation_path):
+            log.info("Site not found %s" % site_path, flush=True)
+            continue
+
+        log.info("Clustering %s" % site_path, flush=True)
+        site_supp_files_folder = os.path.join(supp_folder,
+                                              '%s-supps' % site[:2],
+                                              '%s' % site)
+
+        if os.path.exists(os.path.join(site_supp_files_folder, 'cell_pixel_assignments.pkl')) and not rerun:
+            log.info('Found previously saved instance clustering output in {}. Skip processing...'
+                  .format(site_supp_files_folder))
+            continue
+        elif not os.path.exists(site_supp_files_folder):
+            os.makedirs(site_supp_files_folder, exist_ok=True)
+
+        process_site_instance_segmentation(site_path,
+                                           site_segmentation_path,
+                                           site_supp_files_folder,
+                                           **kwargs)
+    return
+
 
 def extract_patches(raw_folder: str,
                     supp_folder: str,
