+++ pipeline/segmentation.py
@@ -4,7 +4,7 @@ import os
 import numpy as np
 from NNsegmentation.models import Segment
 from NNsegmentation.data import load_input, predict_whole_map
-from SingleCellPatch.instance_clustering import process_site_instance_segmentation
+# from SingleCellPatch.instance_clustering import process_site_instance_segmentation
 from configs.config_reader import YamlReader
 import logging
 log = logging.getLogger(__name__)
@@ -86,56 +86,3 @@ def segmentation(raw_folder_: str,
                 log.error("Error in predicting site %s" % site)
     return
 
-
-def instance_segmentation(raw_folder: str,
-                          supp_folder: str,
-                          val_folder: str,
-                          sites: list,
-                          config_: YamlReader,
-                          rerun=False,
-
-                          **kwargs):
-    """ Helper function for instance segmentation
-
-    Wrapper method `process_site_instance_segmentation` will be called, which
-    loads "*_NNProbabilities.npy" files and performs instance segmentation.
-
-    Results will be saved in the supplementary data folder, including:
-        "cell_positions.pkl": dict of cells in each frame (IDs and positions);
-        "cell_pixel_assignments.pkl": dict of pixel compositions of cells
-            in each frame;
-        "segmentation_*.png": image of instance segmentation results.
-
-    Args:
-        raw_folder (str): folder for raw data, segmentation and summarized results
-        supp_folder (str): folder for supplementary data
-        sites (list of str): list of site names
-        config (YamlReader):
-
-    """
-
-    for site in sites:
-        site_path = os.path.join(raw_folder, '%s.npy' % site)
-        site_segmentation_path = os.path.join(raw_folder,
-                                              '%s_NNProbabilities.npy' % site)
-        if not os.path.exists(site_path) or not os.path.exists(site_segmentation_path):
-            log.info("Site not found %s" % site_path, flush=True)
-            continue
-
-        log.info("Clustering %s" % site_path, flush=True)
-        site_supp_files_folder = os.path.join(supp_folder,
-                                              '%s-supps' % site[:2],
-                                              '%s' % site)
-
-        if os.path.exists(os.path.join(site_supp_files_folder, 'cell_pixel_assignments.pkl')) and not rerun:
-            log.info('Found previously saved instance clustering output in {}. Skip processing...'
-                  .format(site_supp_files_folder))
-            continue
-        elif not os.path.exists(site_supp_files_folder):
-            os.makedirs(site_supp_files_folder, exist_ok=True)
-
-        process_site_instance_segmentation(site_path,
-                                           site_segmentation_path,
-                                           site_supp_files_folder,
-                                           **kwargs)
-    return
